split-app/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
pymongo==4.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
motor==3.3.1
gunicorn==21.2.0
streamlit==1.30.0
requests==2.31.0